Browser support table.
The first (and only) table with browser compatibility was for the data
URI scheme (table full of restrictions and issues).
Now the first thing the user will see is where the library will work :
everywhere.

TODO : test in older browsers. IE and Firefox are good but
Opera/Safari/Chrome lack tests with old versions.