Add support for the unicode path extra field
This patch sets the unicode path extra field. unzip needs at least one
extra field to correctly handle unicode path, so using the path is as
good as any other information. This could improve the situation with
other archive managers too.

This field is usually used without the utf8 flag, with a non unicode
path in the header (winrar, winzip). This helps (a bit) with the messy
Windows' default compressed folders feature but breaks on p7zip which
doesn't seek the unicode path extra field.

So for now, UTF-8 everywhere !

Fix #79.