Rewrite the documentation
This commit changes a lot. Instead of one (very) long page explaining
the API and hidden examples, this commit uses jekyll and gh-pages to
provide multiple pages written in markdown. It also adds other examples,
how-to, etc. Now, most pages are written in markdown. Some are still
written in html : the examples, the first page, the html template (they
need far more html elements than the others).

I used twitter bootstrap for the css but without any customization so the
pages will look a bit generic. The last time I tried to be creative with
css ended horribly so I will let people with more design/ergonomics
skills do that.

Regarding the design / ergonomics of the pages, don't hesitate to voice
your opinion !

I also added a copy of FileSaver.js as it solve the "how to download the
file" in a cleaner and (I think) in a more future proof way.

Last but not least, English is not my mother tongue : don't hesitate to
correct me :)