Add ZipCrypto library for decryption
For ZipCrypto encrypted files, one of the following MUST be passed in the loadOptions object:
* retrievePasswordCallback: a function that returns a password as a string, or
* retrieveEncryptionKeysCallback: a function that returns an array of 3 32-bit integers to be used as pre-primed encryption keys
If both are set, ZipCrypto will call only retrieveEncryptionKeysCallback.

Additionally, invalidPasswordCallback MAY be set and will be called if the supplied password does not validate.
invalidPasswordCallback should return `true` if password validation should be ignored.

This patch does _not_ add support for PKWARE's Strong Encryption.