Improve backward compatibility for ZipObject
If the user changes the metadata after the creation of the ZipObject AND
uses `JSZip#file` to access directly to the ZipObject, he will get a
bug. With this patch, we now store the initial values of dir and date
(changing dir afterward is a bit dangerous but you can't guess what a
user would do).
If the date (or dir) has been updated via the new path, we use the new
value. Otherwise, we use the value stored at the old path. That way, we
don't need to check the old path against the initial value.