Don't include the Buffer polyfill in dist/ files.
In a browser, there are multiple implementations of nodejs' Buffer. If
we let browserify includes its own, we won't play nicely with the others
(it uses `instanceof` and doesn't accept other implementations).
With this patch, we don't include any implementation but we let the
environment provides (or not) the Buffer.