Remove usage of Text{En,De}coder.
The goal was to provide a faster utf8 encoding / decoding. This API is only
available on Firefox and it doesn't work well in a Firefox addon context
(the generated Uint8Array and the available Uint8Array class come from
different contexts, leading to bugs and performance issues).

Instead of adding more conditions to (try to) detect if the Uint8Array
is from the same context or not, I think it's better to remove this
optimization for now.

Fix #151.