Rework the unix/dos permissions and the dir option.
This is an update for #200.

If we don't keep the original unix/dos permissions, this will lead to
hard-to-debug issues, where the permissions are not exactly like they
were specified.
This commit changes that and uses the exact file mode. It also updates the
default umask from 022 to 002 (with unixPermissions == null, files will
get 0664 and folders 0775).

To put a folder with custom permissions, folder() is not a good
candidate : its behavior is to re-use existing entries. Instead, this
commit documents the existing dir attribute.