[breaking change] Remove sync methods, part 1.
This commit removes most of the sync methods (everything but the load
method).
The new getters are :

- zip.async(type) : es6 promises
- zip.stream(type) : nodejs Stream3
- zip.internalStream(type) : our internal stream (useful to create wrappers).

The new generate methods are :
- zip.generateAsync() : es6 promises
- zip.generateStream() : nodejs Stream3
- zip.generateInternalStream() : our internal stream

With the "stream" keyword, everyone will think of the nodejs stream :
it's safer to use "internalStream" for our own.
The old sync versions will throw an exception.

Also fix the stream version we use : we use Stream3, not the stream
offered by the current node runtime.

Also changed in generate : prefer "binarystring" over "string".
That way, JSZip#generate will be consistent with ZipObject#async.
"string" is still supported because a zip as a text don't make sense.