Add decodeFileName in load().
Some archive managers use the machine encoding to generate the file name and
comment. The user can recover the content of a file not encoded in UTF-8
but can't do anything for the file name. This new option adds the
ability to decode file names and comments with a custom algorithm.

See #210.