Add encodeFileName in generate().
This option allows the user to generate a zip file with a custom encoding
for the entry file name / comment. See also "decodeFileName" in load().
Not using UTF-8 when generating a zip file is maybe asking for troubles but
some archive managers (Windows Compressed Folders for example) don't
support well unicode stuff.

See #210.