[breaking change] Mandatory type in generate().
Calling `generate()` won't work anymore, the developer needs to specify an
output type.

"base64" was a nice default when the main way to provide a zip to a user
was to do:

    window.location = "data:application/zip;base64," + zip.generate({type:"base64"});

We now have better alternatives (saveAs, blob url, etc) that aren't
limited by the number of characters you can put in an url.