Add support of Array in JSZip#load.
In #250 case, we don't have fancy Uint8Array but we have a unsupported
binary format. An array of bytes (numbers between 0 and 255) is the
lowest common denominator. A binary string would awkward to build here
and building it reliably can be tricky (without filling the stack or
taking too much time/memory).
This commit adds the missing `arrayReader` needed here.