Nodejs stream and loadAsync: add an error.
A zip file is designed to be read backward so a stream is not suited
for this task. We could read the whole stream in memory but that would
be very counter-intuitive. The previous message is too generic ("The
data of [...] is in an unsupported format") and a better message will
save hours of debugging.