Browserify: add global in the build.
JSZip doesn't need it but `lie` (the `Promise` shim) does. This
dependency is only called when the js runtime doesn't provide a `Promise`.

The minimalist browserify flags (to work in all browsers) are `--no-builtins
--insert-global-vars global` and the minimalist webpack configuration is now

```js
module.exports = {
  node: {
    "stream": "empty",
    global: true,
    process: false,
    Buffer: false
  }
};
```

If the `global` object is not available in the bundle, you will get an
error in old browsers.