More robust type checking
If `input` comes from a different context, it might not be an instance of `window.Uint8Array` or `window.ArrayBuffer`. In particular, when reading a file in a Firefox extension using `OS.File.read`, a `Uint8Array` is returned that does not pass the `x instanceof Uint8Array` test.

See https://stackoverflow.com/questions/15251879/how-to-check-if-a-variable-is-a-typed-array-in-javascript for more details.